Blocks are now sendt properly to clients

+++ Source/Constants.h
@@ -75,7 +75,7 @@ typedef int EntityID;
  * \brief	Block IDs.
  */
 
-typedef int BlockID;
+typedef short BlockID;
 
 /**
  * \enum	BlockEntityID
